[Refactor] rename `codeMirrror.ts` to `codeMirror.ts`
